refactor(apps/frontend-pwa): separate account and avatar changes into separate forms
