+++ apps/frontend-pwa/src/pages/editProfile.tsx
@@ -1,62 +1,22 @@
-import { useMutation, useQuery } from '@apollo/client'
-import { BigHead } from '@bigheads/core'
-import { faSave } from '@fortawesome/free-regular-svg-icons'
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
-import {
-  DeleteParticipantAccountDocument,
-  LogoutParticipantDocument,
-  SelfDocument,
-  UpdateParticipantProfileDocument,
-} from '@klicker-uzh/graphql/dist/ops'
+import { useQuery } from '@apollo/client'
+import AccountDeletionForm from '@components/forms/AccountDeletionForm'
+import AvatarUpdateForm from '@components/forms/AvatarUpdateForm'
+import UpdateAccountInfoForm from '@components/forms/UpdateAccountInfoForm'
+import { SelfDocument } from '@klicker-uzh/graphql/dist/ops'
 import Loader from '@klicker-uzh/shared-components/src/Loader'
-import { AVATAR_OPTIONS } from '@klicker-uzh/shared-components/src/constants'
 import { addApolloState, initializeApollo } from '@lib/apollo'
 import { getParticipantToken } from '@lib/token'
-import {
-  Button,
-  FormikSelectField,
-  FormikSwitchField,
-  FormikTextField,
-  H3,
-  Modal,
-  Prose,
-  Toast,
-} from '@uzh-bf/design-system'
-import { Form, Formik } from 'formik'
+import { Toast } from '@uzh-bf/design-system'
 import { GetServerSidePropsContext } from 'next'
 import { useTranslations } from 'next-intl'
-import hash from 'object-hash'
-import { pick } from 'ramda'
-import { useEffect, useState } from 'react'
-import * as yup from 'yup'
+import { useState } from 'react'
 import Layout from '../components/Layout'
-import DebouncedUsernameField from '../components/forms/DebouncedUsernameField'
 
 function EditProfile() {
   const t = useTranslations()
   const { data, loading } = useQuery(SelfDocument)
-  const [updateParticipantProfile] = useMutation(
-    UpdateParticipantProfileDocument
-  )
-  const [deleteParticipantAccount] = useMutation(
-    DeleteParticipantAccountDocument
-  )
-  const [logoutParticipant] = useMutation(LogoutParticipantDocument)
-
-  const [decodedRedirectPath, setDecodedRedirectPath] = useState('/profile')
   const [showError, setShowError] = useState(false)
-  const [deleteModalOpen, setDeleteModalOpen] = useState(false)
-  const [isUsernameAvailable, setIsUsernameAvailable] = useState<
-    boolean | undefined
-  >(true)
-
-  useEffect(() => {
-    const urlParams = new URLSearchParams(window?.location?.search)
-    const redirectTo = urlParams?.get('redirect_to')
-    if (redirectTo) {
-      setDecodedRedirectPath(decodeURIComponent(redirectTo))
-    }
-  }, [])
+  const [showSuccess, setShowSuccess] = useState(false)
 
   if (loading || !data?.self) {
     return (
@@ -74,321 +34,25 @@ function EditProfile() {
       course={{ displayName: t('shared.generic.title') }}
       displayName={t('pwa.profile.editProfile')}
     >
-      <Formik
-        validationSchema={yup.object({
-          email: yup
-            .string()
-            .required(t('pwa.profile.emailRequired'))
-            .email(t('pwa.profile.emailInvalid')),
-          username: yup
-            .string()
-            .required(t('pwa.profile.usernameRequired'))
-            .min(5, t('pwa.profile.usernameMinLength', { length: '5' }))
-            .max(15, t('pwa.profile.usernameMaxLength', { length: '15' }))
-            .test(
-              'isUsernameAvailable',
-              t('pwa.createAccount.usernameAvailability'),
-              () =>
-                typeof isUsernameAvailable === 'undefined' ||
-                isUsernameAvailable === true
-            ),
-          password: yup
-            .string()
-            .optional()
-            .min(8, t('pwa.profile.passwordMinLength', { length: '8' })),
-          passwordRepetition: yup.string().when('password', {
-            is: (val: string) => val && val.length > 0,
-            then: (schema) =>
-              schema
-                .required(t('pwa.profile.identicalPasswords'))
-                .min(8, t('pwa.profile.passwordMinLength', { length: '8' }))
-                .oneOf(
-                  [yup.ref('password'), null],
-                  t('pwa.profile.identicalPasswords')
-                ),
-            otherwise: (schema) =>
-              schema.oneOf([''], t('pwa.profile.identicalPasswords')),
-          }),
-        })}
-        initialValues={{
-          isProfilePublic: data.self.isProfilePublic,
-          email: data.self.email,
-          username: data.self.username,
-          password: '',
-          passwordRepetition: '',
-
-          // TODO: canton or country on the shirts
-          skinTone:
-            data.self.avatarSettings?.skinTone ?? AVATAR_OPTIONS.skinTone[0],
-          eyes: data.self.avatarSettings?.eyes ?? AVATAR_OPTIONS.eyes[0],
-          mouth: data.self.avatarSettings?.mouth ?? AVATAR_OPTIONS.mouth[0],
-          hair: data.self.avatarSettings?.hair ?? AVATAR_OPTIONS.hair[0],
-          facialHair:
-            data.self.avatarSettings?.facialHair ??
-            AVATAR_OPTIONS.facialHair[0],
-          accessory:
-            data.self.avatarSettings?.accessory ?? AVATAR_OPTIONS.accessory[0],
-          hairColor:
-            data.self.avatarSettings?.hairColor ?? AVATAR_OPTIONS.hairColor[0],
-          clothing:
-            data.self.avatarSettings?.clothing ?? AVATAR_OPTIONS.clothing[0],
-          clothingColor:
-            data.self.avatarSettings?.clothingColor ??
-            AVATAR_OPTIONS.clothingColor[0],
-        }}
-        onSubmit={async (values, { setSubmitting }) => {
-          setSubmitting(true)
-
-          const definition = {
-            skinTone: values.skinTone,
-            eyes: values.eyes,
-            mouth: values.mouth,
-            hair: values.hair,
-            facialHair: values.facialHair,
-            body: 'chest',
-            accessory: values.accessory,
-            hairColor: values.hairColor,
-            clothingColor: values.clothingColor,
-            eyebrows: 'raised',
-            faceMask: false,
-            lashes: false,
-            mask: false,
-            clothing: values.clothing,
-            graphic: 'none',
-            hat: 'none',
-          }
-
-          const avatarHash = hash(definition)
-
-          const result = await updateParticipantProfile({
-            variables: {
-              isProfilePublic: values.isProfilePublic,
-              password: values.password,
-              username: values.username,
-              email: values.email,
-              avatar: avatarHash,
-              avatarSettings: pick(
-                [
-                  'skinTone',
-                  'eyes',
-                  'mouth',
-                  'hair',
-                  'clothingColor',
-                  'clothing',
-                  'accessory',
-                  'hairColor',
-                  'facialHair',
-                ],
-                definition
-              ),
-            },
-          })
-
-          if (result.data?.updateParticipantProfile && !result.errors) {
-            // router.replace(decodedRedirectPath)
-          } else {
-            setShowError(true)
-          }
-        }}
-      >
-        {({ values, isSubmitting, isValid, validateField }) => {
-          return (
-            <Form>
-              <div className="flex flex-col gap-4 md:w-full md:max-w-5xl md:mx-auto">
-                <div className="flex flex-col w-full gap-4 md:flex-row">
-                  <div className="flex flex-col justify-between flex-1 order-2 gap-3 rounded md:order-1 md:bg-slate-50 md:p-4">
-                    <div>
-                      <H3 className={{ root: 'border-b mb-0' }}>
-                        {t('shared.generic.profile')}
-                      </H3>
-                      <div className="mb-2 space-y-3">
-                        <FormikTextField
-                          // TODO: as soon as verification mechanism for email is implemented, add check for "isEmailValid" in DB for disabled field as emails with typos cannot be changed currently
-                          disabled={
-                            data.self?.email !== '' &&
-                            data.self?.email !== null &&
-                            typeof data.self?.email !== 'undefined' &&
-                            !!values.email
-                          }
-                          name="email"
-                          label={t('shared.generic.email')}
-                          labelType="small"
-                          className={{ label: 'font-bold text-md text-black' }}
-                        />
-                        <DebouncedUsernameField
-                          name="username"
-                          label={t('shared.generic.username')}
-                          valid={isUsernameAvailable}
-                          setValid={(usernameAvailable: boolean | undefined) =>
-                            setIsUsernameAvailable(usernameAvailable)
-                          }
-                          validateField={async () => {
-                            await validateField('username')
-                          }}
-                        />
-                        <FormikTextField
-                          name="password"
-                          label={t('shared.generic.password')}
-                          labelType="small"
-                          className={{ label: 'font-bold text-md text-black' }}
-                          type="password"
-                        />
-                        <FormikTextField
-                          name="passwordRepetition"
-                          label={t('shared.generic.passwordRepetition')}
-                          labelType="small"
-                          className={{ label: 'font-bold text-md text-black' }}
-                          type="password"
-                        />
-                        <div>
-                          <div className="font-bold">
-                            {t('pwa.profile.publicProfile')}
-                          </div>
-                          <div className="flex flex-row gap-4 space-between">
-                            <div className="flex flex-col items-center gap-1">
-                              <FormikSwitchField name="isProfilePublic" />
-                              {values.isProfilePublic
-                                ? t('shared.generic.yes')
-                                : t('shared.generic.no')}
-                            </div>
-                            <div className="flex-1">
-                              <Prose className={{ root: 'prose-sm' }}>
-                                {t('pwa.profile.isProfilePublic')}
-                              </Prose>
-                            </div>
-                          </div>
-                        </div>
-                      </div>
-                    </div>
-
-                    <Button
-                      fluid
-                      type="submit"
-                      disabled={isSubmitting || !isValid}
-                    >
-                      <Button.Icon>
-                        <FontAwesomeIcon icon={faSave} />
-                      </Button.Icon>
-                      <Button.Label>{t('shared.generic.save')}</Button.Label>
-                    </Button>
-                  </div>
-
-                  <div className="flex flex-col justify-between flex-1 order-1 space-y-4 rounded md:order-2 md:bg-slate-50 md:p-4">
-                    <div className="flex-initial space-y-2">
-                      <H3 className={{ root: 'border-b mb-0' }}>
-                        {t('pwa.profile.deleteProfile')}
-                      </H3>
-                      <Prose className={{ root: '' }}>
-                        {t('pwa.profile.deleteProfileDescription')}
-                      </Prose>
-
-                      <Modal
-                        title={t('pwa.profile.deleteProfile')}
-                        open={deleteModalOpen}
-                        onClose={(): void => setDeleteModalOpen(false)}
-                        hideCloseButton={true}
-                        className={{ content: 'h-max max-w-md' }}
-                        onPrimaryAction={
-                          <Button
-                            className={{
-                              root: 'bg-red-600 border-red-700 text-white',
-                            }}
-                            onClick={async () => {
-                              await deleteParticipantAccount()
-                              try {
-                                await logoutParticipant()
-                              } catch (e) {}
-                              window?.location.reload()
-                            }}
-                          >
-                            {t('shared.generic.confirm')}
-                          </Button>
-                        }
-                        onSecondaryAction={
-                          <Button onClick={() => setDeleteModalOpen(false)}>
-                            {t('shared.generic.cancel')}
-                          </Button>
-                        }
-                        trigger={
-                          <Button
-                            onClick={(): void => setDeleteModalOpen(true)}
-                          >
-                            {t('shared.generic.delete')}
-                          </Button>
-                        }
-                      >
-                        {t('pwa.profile.deleteProfileConfirmation')}
-                      </Modal>
-                    </div>
-                  </div>
-                </div>
-
-                <div className="space-y-2 rounded md:bg-slate-50 md:p-4">
-                  <H3 className={{ root: 'border-b' }}>Avatar</H3>
-                  <div className="flex flex-col gap-4 md:flex-row md:gap-8">
-                    <div className="flex-1">
-                      <BigHead
-                        // @ts-ignore
-                        className="w-full border-b-4 md:h-48 border-primary-80"
-                        eyebrows="raised"
-                        faceMask={false}
-                        lashes={false}
-                        mask={false}
-                        clothing={values.clothing}
-                        hat="none"
-                        graphic="none"
-                        accessory={values.accessory}
-                        body="chest"
-                        skinTone={values.skinTone}
-                        clothingColor={values.clothingColor}
-                        eyes={values.eyes}
-                        hair={values.hair}
-                        hairColor={values.hairColor}
-                        mouth={values.mouth}
-                        facialHair={values.facialHair}
-                      />
-                    </div>
-                    <div className="grid grid-cols-2 gap-2 text-xs md:grid-cols-3 md:text-sm">
-                      {Object.keys(AVATAR_OPTIONS).map((key) => (
-                        <FormikSelectField
-                          className={{
-                            root: 'w-full',
-                            label: 'font-bold text-md',
-                            select: {
-                              root: 'w-full md:w-36',
-                              trigger: 'w-full md:p-1 justify-between md:px-3',
-                            },
-                          }}
-                          key={key}
-                          label={t(`pwa.avatar.${key}`)}
-                          labelType="small"
-                          name={key}
-                          items={AVATAR_OPTIONS[key].map((value) => {
-                            return {
-                              label: t(`pwa.avatar.${value}`),
-                              value: value,
-                            }
-                          })}
-                        />
-                      ))}
-                    </div>
-                  </div>
-                  <Button
-                    fluid
-                    type="submit"
-                    disabled={isSubmitting || !isValid}
-                  >
-                    <Button.Icon>
-                      <FontAwesomeIcon icon={faSave} />
-                    </Button.Icon>
-                    <Button.Label>{t('shared.generic.save')}</Button.Label>
-                  </Button>
-                </div>
-              </div>
-            </Form>
-          )
-        }}
-      </Formik>
+      <div className="flex flex-col gap-8 md:gap-4 md:w-full md:max-w-5xl md:mx-auto">
+        <div className="flex flex-col w-full gap-8 md:gap-4 md:flex-row">
+          <div className="w-full md:w-1/2">
+            <UpdateAccountInfoForm
+              user={data.self}
+              setShowError={setShowError}
+              setShowSuccess={setShowSuccess}
+            />
+          </div>
+          <div className="w-full md:w-1/2">
+            <AccountDeletionForm />
+          </div>
+        </div>
+        <AvatarUpdateForm
+          user={data.self}
+          setShowError={setShowError}
+          setShowSuccess={setShowSuccess}
+        />
+      </div>
       <Toast
         type="error"
         openExternal={showError}
@@ -396,6 +60,14 @@ function EditProfile() {
         duration={8000}
       >
         {t('pwa.profile.editProfileFailed')}
+      </Toast>
+      <Toast
+        type="success"
+        openExternal={showSuccess}
+        setOpenExternal={setShowSuccess}
+        duration={4000}
+      >
+        {t('pwa.profile.editProfileSuccess')}
       </Toast>
     </Layout>
   )
