+++ packages/graphql/src/public/client.json
@@ -70,7 +70,7 @@
    "UnpublishMicroSession": "baf78cd8b16dd3c9e3ba91202c4133524c2befb2b73338026e16b2364c89aa52",
    "UnsubscribeFromPush": "3bd2659b4e7b705ddaf7168504c8f490520f884f1501f4568889711016da739b",
    "UpdateGroupAverageScores": "fccda9bacb2fb5740670ff2b65ef9258bc41a0ff97e5be4a94f523669fa4f46d",
-   "UpdateParticipantProfile": "f090a995bb8e8fd1639259ae8406aeafed6fd3f543b882e80aecab70dea8b51d",
+   "UpdateParticipantProfile": "60ee432348ff1d4baa08e9e3638a3e5b7396dece85b46291b2e864d088441497",
    "UpdateTagOrdering": "8d865d09b3ef6613f8a4b2ab1c679442f5920e70c89eb9c9456b68f922f499a5",
    "UpvoteFeedback": "1bb1b9c4d54dd21359cb8b10323958fe15776d603b4ab9d54de7c6bb779cad43",
    "VoteFeedbackResponse": "eca425e3d49e7898281d10de07a8b54068e6f4ce77dc5049bd9f2386395fe459",
@@ -124,5 +124,6 @@
    "FeedbackCreated": "1447a6241024f772cef8129c6676933db63ab227a0744e9e2ff73a024d838759",
    "FeedbackRemoved": "483f0c5d7011474da36ec8feb384ec2f9e5877a4283f75f6718da405e616bab6",
    "FeedbackUpdated": "8df9b3d8231f9e1e79dc90873dd080f34b67c7ecb14642babb6c6e482926e912",
-   "RunningSessionUpdated": "4e5e5bfec09446168b6941f64c58d0fa249fd7942f7aa239e3faadb4012f96fd"
+   "RunningSessionUpdated": "4e5e5bfec09446168b6941f64c58d0fa249fd7942f7aa239e3faadb4012f96fd",
+   "UpdateParticipantAvatar": "ea07d0ab75a016dd988627fcd6de74bdba36de21c8c2c6b59e4a8a7488758519"
 }
