+++ packages/graphql/src/public/schema.graphql
@@ -569,7 +569,8 @@ type Mutation {
   unpublishMicroSession(id: String!): MicroSession
   unsubscribeFromPush(courseId: String!, endpoint: String!): Boolean
   updateGroupAverageScores: Boolean!
-  updateParticipantProfile(avatar: String, avatarSettings: AvatarSettingsInput, email: String, isProfilePublic: Boolean, password: String, username: String): Participant
+  updateParticipantAvatar(avatar: String!, avatarSettings: AvatarSettingsInput!): Participant
+  updateParticipantProfile(email: String!, isProfilePublic: Boolean!, password: String!, username: String!): Participant
   updateTagOrdering(originIx: Int!, targetIx: Int!): [Tag!]
   upvoteFeedback(feedbackId: Int!, increment: Int!): Feedback
   voteFeedbackResponse(id: Int!, incrementDownvote: Int!, incrementUpvote: Int!): FeedbackResponse
