+++ packages/graphql/src/public/server.json
@@ -70,7 +70,7 @@
    "baf78cd8b16dd3c9e3ba91202c4133524c2befb2b73338026e16b2364c89aa52": "mutation UnpublishMicroSession($id: String!) {\n  unpublishMicroSession(id: $id) {\n    id\n    name\n    displayName\n    status\n    numOfInstances\n    __typename\n  }\n}",
    "3bd2659b4e7b705ddaf7168504c8f490520f884f1501f4568889711016da739b": "mutation UnsubscribeFromPush($courseId: String!, $endpoint: String!) {\n  unsubscribeFromPush(courseId: $courseId, endpoint: $endpoint)\n}",
    "fccda9bacb2fb5740670ff2b65ef9258bc41a0ff97e5be4a94f523669fa4f46d": "mutation UpdateGroupAverageScores {\n  updateGroupAverageScores\n}",
-   "f090a995bb8e8fd1639259ae8406aeafed6fd3f543b882e80aecab70dea8b51d": "mutation UpdateParticipantProfile($avatar: String, $avatarSettings: AvatarSettingsInput, $username: String, $password: String, $email: String, $isProfilePublic: Boolean) {\n  updateParticipantProfile(avatar: $avatar, avatarSettings: $avatarSettings, username: $username, password: $password, email: $email, isProfilePublic: $isProfilePublic) {\n    id\n    username\n    email\n    isProfilePublic\n    avatar\n    avatarSettings\n    __typename\n  }\n}",
+   "60ee432348ff1d4baa08e9e3638a3e5b7396dece85b46291b2e864d088441497": "mutation UpdateParticipantProfile($username: String!, $password: String!, $email: String!, $isProfilePublic: Boolean!) {\n  updateParticipantProfile(username: $username, password: $password, email: $email, isProfilePublic: $isProfilePublic) {\n    id\n    username\n    email\n    isProfilePublic\n    __typename\n  }\n}",
    "8d865d09b3ef6613f8a4b2ab1c679442f5920e70c89eb9c9456b68f922f499a5": "mutation UpdateTagOrdering($originIx: Int!, $targetIx: Int!) {\n  updateTagOrdering(originIx: $originIx, targetIx: $targetIx) {\n    id\n    order\n    __typename\n  }\n}",
    "1bb1b9c4d54dd21359cb8b10323958fe15776d603b4ab9d54de7c6bb779cad43": "mutation UpvoteFeedback($feedbackId: Int!, $increment: Int!) {\n  upvoteFeedback(feedbackId: $feedbackId, increment: $increment) {\n    id\n    votes\n    __typename\n  }\n}",
    "eca425e3d49e7898281d10de07a8b54068e6f4ce77dc5049bd9f2386395fe459": "mutation VoteFeedbackResponse($id: Int!, $incrementUpvote: Int!, $incrementDownvote: Int!) {\n  voteFeedbackResponse(id: $id, incrementUpvote: $incrementUpvote, incrementDownvote: $incrementDownvote) {\n    id\n    positiveReactions\n    negativeReactions\n    __typename\n  }\n}",
@@ -124,5 +124,6 @@
    "1447a6241024f772cef8129c6676933db63ab227a0744e9e2ff73a024d838759": "fragment FeedbackData on Feedback {\n  id\n  isPublished\n  isPinned\n  isResolved\n  content\n  votes\n  resolvedAt\n  createdAt\n  responses {\n    id\n    content\n    positiveReactions\n    negativeReactions\n    __typename\n  }\n  __typename\n}\nsubscription FeedbackCreated($sessionId: String!) {\n  feedbackCreated(sessionId: $sessionId) {\n    ...FeedbackData\n    __typename\n  }\n}",
    "483f0c5d7011474da36ec8feb384ec2f9e5877a4283f75f6718da405e616bab6": "subscription FeedbackRemoved($sessionId: String!) {\n  feedbackRemoved(sessionId: $sessionId)\n}",
    "8df9b3d8231f9e1e79dc90873dd080f34b67c7ecb14642babb6c6e482926e912": "fragment FeedbackData on Feedback {\n  id\n  isPublished\n  isPinned\n  isResolved\n  content\n  votes\n  resolvedAt\n  createdAt\n  responses {\n    id\n    content\n    positiveReactions\n    negativeReactions\n    __typename\n  }\n  __typename\n}\nsubscription FeedbackUpdated($sessionId: String!) {\n  feedbackUpdated(sessionId: $sessionId) {\n    ...FeedbackData\n    __typename\n  }\n}",
-   "4e5e5bfec09446168b6941f64c58d0fa249fd7942f7aa239e3faadb4012f96fd": "fragment QuestionDataWithoutSolutions on QuestionInstance {\n  questionData {\n    id\n    name\n    type\n    displayMode\n    content\n    explanation\n    pointsMultiplier\n    ... on ChoicesQuestionData {\n      options {\n        choices {\n          ix\n          value\n          __typename\n        }\n        __typename\n      }\n      __typename\n    }\n    ... on NumericalQuestionData {\n      options {\n        unit\n        accuracy\n        placeholder\n        restrictions {\n          min\n          max\n          __typename\n        }\n        __typename\n      }\n      __typename\n    }\n    ... on FreeTextQuestionData {\n      options {\n        restrictions {\n          maxLength\n          __typename\n        }\n        __typename\n      }\n      __typename\n    }\n    __typename\n  }\n  __typename\n}\nsubscription RunningSessionUpdated($sessionId: String!) {\n  runningSessionUpdated(sessionId: $sessionId) {\n    id\n    status\n    expiresAt\n    timeLimit\n    randomSelection\n    execution\n    instances {\n      id\n      ...QuestionDataWithoutSolutions\n      __typename\n    }\n    __typename\n  }\n}"
+   "4e5e5bfec09446168b6941f64c58d0fa249fd7942f7aa239e3faadb4012f96fd": "fragment QuestionDataWithoutSolutions on QuestionInstance {\n  questionData {\n    id\n    name\n    type\n    displayMode\n    content\n    explanation\n    pointsMultiplier\n    ... on ChoicesQuestionData {\n      options {\n        choices {\n          ix\n          value\n          __typename\n        }\n        __typename\n      }\n      __typename\n    }\n    ... on NumericalQuestionData {\n      options {\n        unit\n        accuracy\n        placeholder\n        restrictions {\n          min\n          max\n          __typename\n        }\n        __typename\n      }\n      __typename\n    }\n    ... on FreeTextQuestionData {\n      options {\n        restrictions {\n          maxLength\n          __typename\n        }\n        __typename\n      }\n      __typename\n    }\n    __typename\n  }\n  __typename\n}\nsubscription RunningSessionUpdated($sessionId: String!) {\n  runningSessionUpdated(sessionId: $sessionId) {\n    id\n    status\n    expiresAt\n    timeLimit\n    randomSelection\n    execution\n    instances {\n      id\n      ...QuestionDataWithoutSolutions\n      __typename\n    }\n    __typename\n  }\n}",
+   "ea07d0ab75a016dd988627fcd6de74bdba36de21c8c2c6b59e4a8a7488758519": "mutation UpdateParticipantAvatar($avatar: String!, $avatarSettings: AvatarSettingsInput!) {\n  updateParticipantAvatar(avatar: $avatar, avatarSettings: $avatarSettings) {\n    avatar\n    avatarSettings\n    __typename\n  }\n}"
 }
